backend/requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
mysqlclient==2.2.1
pymysql==1.1.2  # Alternative MySQL driver for Windows

# Authentication & Security
PyJWT==2.8.0
django-cors-headers==4.3.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1
factory-boy==3.3.0
faker==20.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
qrcode==7.4.2
requests==2.31.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0
